Move `AdBlockingRecoverySetupCTANotice` to settings folder.
